Move util funcs to separate utils file
