--- staging/kos/pkg/controllers/connectionagent/util_types.go
@@ -1,57 +1,0 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package connectionagent
-
-import (
-	"strings"
-
-	k8smetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8sfields "k8s.io/apimachinery/pkg/fields"
-	kosinternalifcs "k8s.io/examples/staging/kos/pkg/client/informers/externalversions/internalinterfaces"
-)
-
-type SliceOfString []string
-
-func (x SliceOfString) Equal(y SliceOfString) bool {
-	if len(x) != len(y) {
-		return false
-	}
-	for i, xi := range x {
-		if xi != y[i] {
-			return false
-		}
-	}
-	return true
-}
-
-type fieldsSelector struct {
-	k8sfields.Selector
-}
-
-func (fs fieldsSelector) toTweakListOptionsFunc() kosinternalifcs.TweakListOptionsFunc {
-	return func(options *k8smetav1.ListOptions) {
-		optionsFieldSelector := options.FieldSelector
-
-		allSelectors := make([]string, 0, 2)
-		if strings.Trim(optionsFieldSelector, " ") != "" {
-			allSelectors = append(allSelectors, optionsFieldSelector)
-		}
-		allSelectors = append(allSelectors, fs.String())
-
-		options.FieldSelector = strings.Join(allSelectors, ",")
-	}
-}
